fix: add Dockerfile with php8.3+mysqli, drop FrankenPHP

+++ config/db.php
@@ -1,12 +1,11 @@
 <?php
-// Railway injects MYSQL_* environment variables automatically
 $host = getenv('MYSQLHOST')     ?: getenv('DB_HOST')     ?: 'localhost';
 $user = getenv('MYSQLUSER')     ?: getenv('DB_USER')     ?: 'root';
 $pass = getenv('MYSQLPASSWORD') ?: getenv('DB_PASSWORD') ?: '';
 $name = getenv('MYSQLDATABASE') ?: getenv('DB_NAME')     ?: 'bloodbank';
-$port = getenv('MYSQLPORT')     ?: 3306;
+$port = (int)(getenv('MYSQLPORT') ?: 3306);
 
-$conn = mysqli_connect($host, $user, $pass, $name, (int)$port);
+$conn = mysqli_connect($host, $user, $pass, $name, $port);
 if (!$conn) {
     die("DB Connection Failed: " . mysqli_connect_error());
 }
